Add gRPC support for discount service integration
- Updated `Basket.API.csproj` to include `Grpc.AspNetCore` and added Protobuf for `discount.proto`.
- Renamed `StoreBasketCommand` to `StoreBasketRequest` in `StoreBasketEndpoints.cs`.
- Modified `StoreBasketHandler.cs` to use `DiscountProtoService.DiscountProtoServiceClient` and added discount deduction logic.
- Changed `Price` type from `decimal` to `double` in `ShoppingCartItem.cs`.
- Configured gRPC client and health checks in `Program.cs`, and added `GrpcSettings` to `appsettings.json`.
- Updated `docker-compose.override.yml` to include gRPC service dependencies.

Branch: feature/grpc-discount-integration

+++ src/Services/Basket/Basket.API/Models/ShoppingCartItem.cs
@@ -6,6 +6,6 @@
         public string Color { get; set; } = default!;
         public string ProductName { get; set; } = default!;
         public Guid ProductId { get; set; } = default!;
-        public decimal Price { get; set; } = default!;
+        public double Price { get; set; } = default!;
     }
 }
